refactor(packages/explorer/src/views/Mining/instructions.js packages/merkle-miner/src/worker.js): Mo
Moved instructions function to external file so that main file is more modular
